refactor email provider helpers
